refactor: split database files
